pyproject.toml
[project]
name = "django-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "factory-boy>=3.3.3",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]
